+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@
 require "rails_psql_jsonb"
 require "logger"
 require "active_record"
+require "database_cleaner/active_record"
 
 ActiveRecord::Base.establish_connection(
   adapter: "postgresql",
@@ -20,11 +21,14 @@ ActiveRecord::Schema.define do
   create_table :friends do |t|
     t.text :name
     t.jsonb :props, default: {}
+    t.jsonb :metadata, default: {}
+    t.timestamps
   end
 
   create_table :mutate_test_friends do |t|
     t.text :name
     t.jsonb :props, default: {}
+    t.timestamps
   end
 end
 
@@ -41,14 +45,20 @@ class MutateTestFriend < ActiveRecord::Base
 end
 
 RSpec.configure do |config|
-  # Enable flags like --only-failures and --next-failure
   config.example_status_persistence_file_path = ".rspec_status"
 
-  # Disable RSpec exposing methods globally on `Module` and `main`
   config.disable_monkey_patching!
 
   config.expect_with :rspec do |c|
     c.syntax = :expect
+  end
+
+  config.before(:suite) do
+    DatabaseCleaner.strategy = :truncation
+  end
+
+  config.around(:each) do |example|
+    DatabaseCleaner.cleaning { example.run }
   end
 
   config.before do